+++ paging/src/lib.rs
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! This crate provides page table–related functions and data structures.
+
+#![no_std]
+#![cfg_attr(verus_only, verifier::allow(unknown_automatic_derive))]
+#![cfg_attr(
+    verus_only,
+    allow(macro_expanded_macro_exports_accessed_by_absolute_paths)
+)]
+
+pub mod address;
+pub mod sizes;
+pub mod util;
