+++ paging/src/specs/align.rs
@@ -4,7 +4,9 @@
 //
 // Author: Ziqiao Zhou <ziqiaozhou@microsoft.com>
 //
-// Specifications related to util.rs that are used in proof_align_down and proof_align_up.
+// Specifications related to alignment helpers that are used in proof_align_down and proof_align_up.
+use core::ops::{Add, BitAnd, Not, Sub};
+use verus_stub::*;
 use vstd::std_specs::ops::{AddSpec, BitAndSpec, NotSpec, SubSpec};
 
 #[verus_verify]
@@ -154,4 +156,7 @@ pub open spec fn spec_is_aligned<T>(val: T, align: T) -> bool where T: IsAligned
 }
 
 } // verus!
-include!("util.proof.verus.rs");
+#[path = "../proofs/align.rs"]
+mod align_proofs;
+
+pub use align_proofs::*;
